test(chat): split read-row link tests into CollapsedActionRows.test.tsx
Keeps CollapsedActions.test.tsx under the 600-line repo-shape cap and
matches the component-per-test-file convention.
