+++ apps/packages/product-client/src/components/workspace/chat/tool-calls/CollapsedActions.test.tsx
@@ -348,109 +348,6 @@ describe("CollapsedActions", () => {
     expect(html).not.toContain("pl-4");
   });
 
-  it("renders read ledger rows as clickable file references", () => {
-    const transcript = createTranscriptState("session-1");
-    transcript.itemsById = {
-      read: toolItem("read", "turn-1", 1, "file_read"),
-    };
-
-    render(
-      <CollapsedActions
-        itemIds={["read"]}
-        transcript={transcript}
-      />,
-    );
-    fireEvent.click(screen.getByRole("button", { name: /Read files/i }));
-
-    const readRow = screen.getByText("Read").parentElement;
-    const badge = readRow?.querySelector("[data-file-reference-badge='inline']");
-    expect(badge?.textContent).toContain("read.ts");
-    expect(badge?.className).toContain("decoration-dotted");
-    expect(badge?.className).toContain("[&>span:first-child]:hidden");
-    expect(readRow?.getAttribute("title")).toContain("read.ts");
-
-    fireEvent.click(badge as Element);
-    expect(openPrimaryMock).toHaveBeenCalledTimes(1);
-    expect(fileReferenceActionsCalls.find((call) => call.rawPath === "read.ts")?.workspacePath)
-      .toBe("read.ts");
-  });
-
-  it("opens raw-input fallback reads through workspace-root inference", () => {
-    const transcript = createTranscriptState("session-1");
-    transcript.itemsById = {
-      read: {
-        ...toolItem("read", "turn-1", 1, "file_read"),
-        contentParts: [],
-        rawInput: { file_path: "src/deep/notes.md" },
-      },
-    };
-
-    render(
-      <CollapsedActions
-        itemIds={["read"]}
-        transcript={transcript}
-      />,
-    );
-    fireEvent.click(screen.getByRole("button", { name: /Read/i }));
-
-    const call = fileReferenceActionsCalls.find((entry) => entry.rawPath === "src/deep/notes.md");
-    expect(call).toBeTruthy();
-    // undefined (not null) so the resolver may infer a workspace-relative path
-    // and the click can open the in-app viewer.
-    expect(call?.workspacePath).toBeUndefined();
-
-    fireEvent.click(screen.getByText("notes.md"));
-    expect(openPrimaryMock).toHaveBeenCalledTimes(1);
-  });
-
-  it("opens parsed shell reads through workspace-root inference", () => {
-    const transcript = createTranscriptState("session-1");
-    transcript.itemsById = {
-      parsed: parsedCommandItem("parsed", "turn-1", 1, [
-        { type: "read", cmd: "cat src/lib/util.ts", path: "src/lib/util.ts", name: "util.ts" },
-      ], "completed"),
-    };
-
-    render(
-      <CollapsedActions
-        itemIds={["parsed"]}
-        transcript={transcript}
-      />,
-    );
-    fireEvent.click(screen.getByRole("button", { name: /Read/i }));
-
-    const call = fileReferenceActionsCalls.find((entry) => entry.rawPath === "src/lib/util.ts");
-    expect(call).toBeTruthy();
-    expect(call?.workspacePath).toBeUndefined();
-
-    fireEvent.click(screen.getByText("util.ts"));
-    expect(openPrimaryMock).toHaveBeenCalledTimes(1);
-  });
-
-  it("keeps structured reads without a workspace path authoritatively external", () => {
-    const transcript = createTranscriptState("session-1");
-    const read = toolItem("read", "turn-1", 1, "file_read");
-    const part = read.contentParts[0];
-    if (part?.type === "file_read") {
-      part.path = "/etc/hosts";
-      part.basename = "hosts";
-      part.workspacePath = null;
-    }
-    transcript.itemsById = { read };
-
-    render(
-      <CollapsedActions
-        itemIds={["read"]}
-        transcript={transcript}
-      />,
-    );
-    fireEvent.click(screen.getByRole("button", { name: /Read/i }));
-
-    const call = fileReferenceActionsCalls.find((entry) => entry.rawPath === "/etc/hosts");
-    expect(call).toBeTruthy();
-    expect(call?.workspacePath).toBeNull();
-  });
-
   it("uses the same Codex ink for command and plain ledger rows", () => {
     const transcript = createTranscriptState("session-1");
     transcript.itemsById = {
